refactor: Update paths in tsconfig.json and import statements
- Update paths in tsconfig.json to include "~/src/*" for "~/*" alias
- Update import statements in various files to use "~/components" and "~/api" aliases instead of relative paths
