+++ _codux/board-templates/1-component-with-data.board.tsx
