+++ _codux/board-templates/2-page-fake-data.board.tsx
