+++ _codux/board-templates/4-component-empty.board.tsx
