+++ _codux/board-templates/5-documentation/new-board.board.tsx
