+++ _codux/boards/components/footer/footer.board.tsx
