+++ _codux/boards/components/hero-image/hero-image.board.tsx
