+++ _codux/boards/customer-details-preview/customer-details-preview.board.tsx
