+++ _codux/boards/lesson-item/lesson-item-mobile-view.board.tsx
