+++ _codux/boards/pages/lesson-page/lesson-page.board.tsx
@@ -1,12 +1,11 @@
-import { LessonPage } from '../../../../pages/lesson-page/lesson-page';
+import { LessonPage } from '~pages/lesson-page/lesson-page';
 import { ContentSlot, createBoard } from '@wixc3/react-board';
-import { PageWrapperRealData } from '/src/_codux/board-wrappers/page-wrapper-real-data';
+import { PageWrapperRealData } from '../../../board-wrappers/page-wrapper-real-data';
 
 export default createBoard({
     name: '8 - LessonPage',
     Board: () => (
         //in practice PageWrapperRealData with a path will render the correct page, but it is less convenient to use in a board
-
         <PageWrapperRealData path="/lesson/yin-yoga">
             <ContentSlot>
                 <LessonPage />
