+++ _codux/boards/pages/lessons-page/page-lessons.board.tsx
@@ -1,7 +1,6 @@
-import { LessonsPage } from '../../../../pages/lessons-page/lessons-page';
+import { LessonsPage } from '~pages/lessons-page/lessons-page';
 import { ContentSlot, createBoard } from '@wixc3/react-board';
-import { PageWrapperRealData } from '/src/_codux/board-wrappers/page-wrapper-real-data';
-import { waitForProductImage } from '/src/_codux/board-wrappers/wait-for-helpers';
+import { PageWrapperRealData } from '../../../board-wrappers/page-wrapper-real-data';
 
 export default createBoard({
     name: '2 - Lessons Page',
@@ -14,8 +13,4 @@ export default createBoard({
         </PageWrapperRealData>
     ),
     isSnippet: false,
-    environmentProps: {
-        canvasMargin: { right: 0, bottom: 0, left: 0 },
-    },
-    readyToSnapshot: waitForProductImage,
 });
