+++ _codux/boards/pages/lessons-page/test-lessons-page-mobile.board.tsx
