+++ _codux/component-templates/default/new-component.module.scss
