+++ _codux/component-templates/default/new-component.tsx
