+++ _codux/component-templates/grid/new-component.module.scss
