+++ _codux/component-templates/grid/new-component.tsx
