+++ _codux/component-templates/stack/new-component.module.scss
