+++ _codux/component-templates/stack/new-component.tsx
